Create metrics file
