+++ multinode/node.go
@@ -9,8 +9,6 @@ import (
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/services"
 )
@@ -24,21 +22,6 @@ var (
 	promPoolRPCNodeVerifiesFailed  *prometheus.CounterVec
 	promPoolRPCNodeVerifiesSuccess *prometheus.CounterVec
 )
-
-func init() {
-	promPoolRPCNodeVerifies = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "pool_rpc_node_verifies",
-		Help: "The total number of chain ID verifications for the given RPC node",
-	}, []string{"network", "chainID", "nodeName"})
-	promPoolRPCNodeVerifiesFailed = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "pool_rpc_node_verifies_failed",
-		Help: "The total number of failed chain ID verifications for the given RPC node",
-	}, []string{"network", "chainID", "nodeName"})
-	promPoolRPCNodeVerifiesSuccess = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "pool_rpc_node_verifies_success",
-		Help: "The total number of successful chain ID verifications for the given RPC node",
-	}, []string{"network", "chainID", "nodeName"})
-}
 
 type NodeConfig interface {
 	PollFailureThreshold() uint32
