+++ multinode/transaction_sender.go
@@ -9,8 +9,6 @@ import (
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/services"
 )
@@ -19,13 +17,6 @@ var (
 	// PromMultiNodeInvariantViolations reports violation of our assumptions
 	PromMultiNodeInvariantViolations *prometheus.CounterVec
 )
-
-func init() {
-	PromMultiNodeInvariantViolations = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "multi_node_invariant_violations",
-		Help: "The number of invariant violations",
-	}, []string{"network", "chainId", "invariant"})
-}
 
 type SendTxResult interface {
 	Code() SendTxReturnCode
